@@ -16,7 +16,7 @@
 
 package com.massivcode.androidmusicplayer.managers;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import com.massivcode.androidmusicplayer.fragments.ArtistFragment;
 import com.massivcode.androidmusicplayer.fragments.PlayerFragment;
